Bug 1308170 - Project-Branch `Date` should build fennec nightly with alternate branding to allow sideloading. r=mshal
DO NOT LAND ON CENTRAL (or backout first)

On the `Date` Project branch we are working on generating Android Nightlies in taskcluster.

To facilitate early testing we should change the android package name to something unique
so it doesn't conflict with actual Fennec nightlies. Especially since early on we will be
using a different signing certificate than the official nightly one (making installing this
harder and potentially profile-losing)

MozReview-Commit-ID: 6QBzCYHdZc