Accumulate dirty rects on displayports so we don't build the displaylist for the entire displayport.
When accumulated the dirty rect from the list of modified frames we stop at any display ports and accumulate the rect on the scroll frame that has a displayport.

When display list building we use this cached rect.

This re-uses the mHasOverrideDirtyRegion bit on the frame. The bit now means we have either a stacking context dirty rect or a scroll frame dirty rect. The presence of absence of the corresponding frame property determines which (or both) the frame has.

Doesn't quite pass the retained-dl reftests with async pan zoom enabled yet. The first reason is that the apz-flush that is done at the end of a reftest invalidates the whole page, which means we build a displaylist for elements we don't want a displaylist for again. If we fix that one test fails with a small difference, perhaps fuzzable, I haven't looked into it.