Invalidate items that might draw images but have no existing geometry when sync decoding
When we've drawn an item previously and have retained geometry, we can use this to determine if the items needs to be redrawn in the presence of sync decoding.
When we don't have existing geometry (possibly because the item had empty bounds and wasn't drawn), then we can't use this, so we instead use the display
item flags to determine if an item *might* draw an image, and invalidate it if so.