remove extraneous task from push_firefox.
The problem here is we assumed that a superset of tasks would exist in the on-push decision task's full-task-graph.json. However, in the promote_firefox action, we add tasks related to partials etc, that didn't exist in the first full-task-graph. That means we haven't been optimizing them out.

The solution here is to download and combine all the full-task-graph.json's, in order, so the rightmost graph will take precedence if they have overlapping labels.