fix fennec notify deps, part ii
I think we were missing post-beetmover-dummy for fennec in general, which is why the notify task happened so quickly. I also don't think we need a ship-phase post-beetmover-dummy set of tasks, because beetmover-cdns pretty much copies everything in a single task.

If this works, we may want to go back and split firefox- and devedition- post-beetmover-dummy into per-build-platform tasks, so we can specify only depending on, say, the win64 post-beetmover-dummy tasks inside of a downstream task's transforms.

I also suspect we don't need post-beetmover-checksums-dummy at all; it's redundant.