Bug 1537764 - P1: Use new function ObserveActivityWithChannelId to inform the observers, r=dragana
Summary:
In socket process, the reference of http channel is not accessiable. So, we use the channel id instead.
There is a table saves the mapping of channel id and the weak pointer of the channel in http handler. We use this mapping to get the real http channel.

--HG--
extra : rebase_source : 6ec97ea2cdacf95231e6403e80038a2fa44837f1